backend/requirements.txt
fastapi==0.111.0
mangum==0.17.0
pydantic==2.7.1
pydantic-settings==2.3.0
boto3==1.34.101
python-jose[cryptography]==3.3.0
httpx==0.27.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
